implemented integration test CommentControllerSpecification

+++ src/main/groovy/com/github/guramkankava/service/MongoDBPostCommentService.groovy
@@ -1,6 +1,7 @@
 package com.github.guramkankava.service
 
 import com.github.guramkankava.document.Comment
+import com.github.guramkankava.exception.ForbiddenException
 import com.github.guramkankava.repository.CommentRepository
 import com.github.guramkankava.repository.PostRepository
 
@@ -17,14 +18,25 @@ class MongoDBPostCommentService implements PostCommentService {
     }
 
     @Override
-    void commentOnPost(String postId, Comment comment) {
+    Comment commentOnPost(String postId, Comment comment) {
         comment.setUsername(authService.getAuthentication().getName())
         comment =  commentRepository.save(comment)
         postRepository.addAComment(postId, comment)
+        comment
     }
 
     @Override
     void deleteAComment(String postId, String commentId) {
-        postRepository.deleteAComment(postId, commentId)
+        if(canDelete(postId, commentId)) {
+            postRepository.deleteAComment(postId, commentId)
+        } else {
+            throw new ForbiddenException('Comment can be deleted by Post or Comment submitter')
+        }
     }
+
+    private boolean canDelete(String postId, String commentId) {
+        String username = authService.getAuthentication().getName()
+        (commentRepository.findByIdAndUsername(commentId, username).isPresent() || postRepository.findByIdAndUsername(postId, username).isPresent())
+    }
+
 }
